Fix dotnet publish compilation errors in Application services
Resolve all CS compiler errors preventing dotnet publish from succeeding:

- Contact: replace non-existent Company/AvatarUrl with CompanyName; use positional record constructors for ContactDto
- Campaign: replace IsEnabled→Enabled; cast int↔CampaignType enum; parse ScheduledAt string to DateTime?
- Portal: replace IsArchived→Archived on entity references (PortalDto retains IsArchived)
- Message: remove non-existent InboxId/IsPrivate→Private; cast int→MessageType enum; map Attachment fields to available entity properties (FallbackTitle/ExternalUrl); use positional MessageDto constructor
- AutomationRule: replace IsActive→Active; default ConditionOperator to "AND"
- Macro: convert Visibility int↔string between MacroDto and entity; handle nullable CreatedById
- Repository: cast long→int for GetByIdAsync calls throughout Conversation/Message/Email services
- CountAsync: replace anonymous-object calls (wrong overload) with .Count on already-fetched collections
- ConversationService: cast between Application.ConversationStatus and Core.Enums.ConversationStatus; implement MuteAsync via Muted property; fix TogglePriorityAsync enum assignment; use positional ConversationDto constructor
- FilterService: cast Core↔Application ConversationStatus to resolve ambiguity from dual enum definitions
- NotificationService: use positional NotificationDto constructor; remove non-existent AccountId/UserId/ReadAt properties

+++ src/CustomerEngagement.Application/Services/Notifications/EmailNotificationService.cs
@@ -46,7 +46,7 @@ public class EmailNotificationService : IEmailNotificationService
         if (user is null || string.IsNullOrEmpty(user.Email))
             return;
 
-        var conversation = await _conversationRepository.GetByIdAsync(conversationId, cancellationToken);
+        var conversation = await _conversationRepository.GetByIdAsync((int)conversationId, cancellationToken);
         if (conversation is null)
             return;
 
@@ -70,7 +70,7 @@ public class EmailNotificationService : IEmailNotificationService
         if (user is null || string.IsNullOrEmpty(user.Email))
             return;
 
-        var message = await _messageRepository.GetByIdAsync(messageId, cancellationToken);
+        var message = await _messageRepository.GetByIdAsync((int)messageId, cancellationToken);
         if (message is null)
             return;
